requirements.txt
getgauge==0.3.18
requests==2.31.0
selenium==4.16.0
webdriver-manager==4.0.1
pytest==7.4.3
appium-python-client==3.1.0
pytest-html==4.1.1
Pillow==10.1.0
allure-pytest==2.13.2
python-dotenv==1.0.0
